Refactor UI entry script into module entrypoint
