+++ src/dndcs/ui/static/index.html
@@ -171,6 +171,6 @@
   </footer>
 
   <dialog id="toast"></dialog>
-  <script src="./app.js"></script>
+  <script type="module" src="./js/main.js"></script>
 </body>
 </html>
